Refactor how models store state
Resources no longer store quantity&produced, now they store produced&consumed,
and make quantity available as a getter.  Also quantity is now a setter, with
logic to maintain the invariants.

The relationship between resources and items has been refactored to be
WriteSelfReadOther (WSRO).  That is, when an item is purchased, it writes the
change to itself, but does not write into the relevant resource, which must
on-demand read the item's state.  This makes the dependency graph sliiightly
less intuitive, but hopefully makes it easier to reason about complex
interactions.

A few other properties needed to be changed to getters to facilitate this
(mostly to cause expression evaluation to happen on-demand instead of up-front).